+++ js/index.js
@@ -391,7 +391,7 @@ class SaveHandler {
         switch (type) {
             case "background":                  console.log("to implement"); break;
             case "image":                       console.log("to implement"); break;
-            case "particle_flow":               console.log("to implement"); break;
+            case "particle_flow":               obj = new imports.visual_objects.VParticleFlow(this, tab.objects, obj_id); break;
             case "text":                        obj = new imports.visual_objects.VText(this, tab.objects, obj_id); break;
             case "timer_straight_bar":          obj = new imports.visual_objects.VTimerStraightBar(this, tab.objects, obj_id); break;
             case "timer_straight_line_point":   obj = new imports.visual_objects.VTimerStraightLinePoint(this, tab.objects, obj_id); break;
@@ -553,6 +553,8 @@ class Project {
         this._save_handler = save_handler;
         this._save_handler.owner_project = this;
     }
+
+    get volume() {return this._volume;}
 
     /*
     #########